Generic Lazy AVRO wrapping
